Add block submission to remote relay
