+++ builder/relay.go
@@ -1,10 +1,9 @@
 package builder
 
 import (
-	"encoding/json"
+	"context"
 	"errors"
 	"fmt"
-	"io"
 	"net/http"
 	"strconv"
 	"strings"
@@ -14,133 +13,20 @@ import (
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/log"
 	boostTypes "github.com/flashbots/go-boost-utils/types"
+	"github.com/flashbots/mev-boost/server"
 )
 
 type testRelay struct {
 	validator ValidatorData
 }
 
+func (r *testRelay) SubmitBlock(msg *BuilderSubmitBlockRequest) error {
+	return nil
+}
 func (r *testRelay) GetValidatorForSlot(nextSlot uint64) (ValidatorData, error) {
 	return r.validator, nil
 }
-func (r *testRelay) GetValidatorsStats() string {
-	return ""
-}
 func (r *testRelay) handleRegisterValidator(w http.ResponseWriter, req *http.Request) {
-}
-
-type LocalRelay struct {
-	beaconClient IBeaconClient
-
-	builderSigningDomain boostTypes.Domain
-
-	validatorsLock sync.RWMutex
-	validators     map[PubkeyHex]ValidatorData
-}
-
-func NewLocalRelay(beaconClient IBeaconClient, builderSigningDomain boostTypes.Domain) *LocalRelay {
-	return &LocalRelay{
-		beaconClient:         beaconClient,
-		builderSigningDomain: builderSigningDomain,
-		validators:           make(map[PubkeyHex]ValidatorData),
-	}
-}
-
-func (r *LocalRelay) handleRegisterValidator(w http.ResponseWriter, req *http.Request) {
-	payload := []boostTypes.SignedValidatorRegistration{}
-	if err := json.NewDecoder(req.Body).Decode(&payload); err != nil {
-		log.Error("could not decode payload", "err", err)
-		respondError(w, http.StatusBadRequest, "invalid payload")
-		return
-	}
-
-	for _, registerRequest := range payload {
-		if len(registerRequest.Message.Pubkey) != 48 {
-			respondError(w, http.StatusBadRequest, "invalid pubkey")
-			return
-		}
-
-		if len(registerRequest.Signature) != 96 {
-			respondError(w, http.StatusBadRequest, "invalid signature")
-			return
-		}
-
-		ok, err := boostTypes.VerifySignature(registerRequest.Message, r.builderSigningDomain, registerRequest.Message.Pubkey[:], registerRequest.Signature[:])
-		if !ok || err != nil {
-			log.Error("error verifying signature", "err", err)
-			respondError(w, http.StatusBadRequest, "invalid signature")
-			return
-		}
-
-		// Do not check timestamp before signature, as it would leak validator data
-		if registerRequest.Message.Timestamp > uint64(time.Now().Add(10*time.Second).Unix()) {
-			respondError(w, http.StatusBadRequest, "invalid payload")
-			return
-		}
-	}
-
-	for _, registerRequest := range payload {
-		pubkeyHex := PubkeyHex(registerRequest.Message.Pubkey.String())
-		if !r.beaconClient.isValidator(pubkeyHex) {
-			respondError(w, http.StatusBadRequest, "not a validator")
-			return
-		}
-	}
-
-	r.validatorsLock.Lock()
-	defer r.validatorsLock.Unlock()
-
-	for _, registerRequest := range payload {
-		pubkeyHex := PubkeyHex(registerRequest.Message.Pubkey.String())
-		if previousValidatorData, ok := r.validators[pubkeyHex]; ok {
-			if registerRequest.Message.Timestamp < previousValidatorData.Timestamp {
-				respondError(w, http.StatusBadRequest, "invalid timestamp")
-				return
-			}
-
-			if registerRequest.Message.Timestamp == previousValidatorData.Timestamp && (registerRequest.Message.FeeRecipient != previousValidatorData.FeeRecipient || registerRequest.Message.GasLimit != previousValidatorData.GasLimit) {
-				respondError(w, http.StatusBadRequest, "invalid timestamp")
-				return
-			}
-		}
-	}
-
-	for _, registerRequest := range payload {
-		pubkeyHex := PubkeyHex(strings.ToLower(registerRequest.Message.Pubkey.String()))
-		r.validators[pubkeyHex] = ValidatorData{
-			Pubkey:       pubkeyHex,
-			FeeRecipient: registerRequest.Message.FeeRecipient,
-			GasLimit:     registerRequest.Message.GasLimit,
-			Timestamp:    registerRequest.Message.Timestamp,
-		}
-
-		log.Info("registered validator", "pubkey", pubkeyHex, "data", r.validators[pubkeyHex])
-	}
-
-	w.WriteHeader(http.StatusOK)
-}
-
-func (r *LocalRelay) GetValidatorForSlot(nextSlot uint64) (ValidatorData, error) {
-	pubkeyHex, err := r.beaconClient.getProposerForNextSlot(nextSlot)
-	if err != nil {
-		return ValidatorData{}, err
-	}
-
-	r.validatorsLock.RLock()
-	if vd, ok := r.validators[pubkeyHex]; ok {
-		r.validatorsLock.RUnlock()
-		return vd, nil
-	}
-	r.validatorsLock.RUnlock()
-	log.Info("no local entry for validator", "validator", pubkeyHex)
-	return ValidatorData{}, errors.New("missing validator")
-}
-
-func (r *LocalRelay) GetValidatorsStats() string {
-	r.validatorsLock.RLock()
-	noValidators := len(r.validators)
-	r.validatorsLock.RUnlock()
-	return fmt.Sprint(noValidators) + " validators registered"
 }
 
 type RemoteRelay struct {
@@ -151,6 +37,7 @@ type RemoteRelay struct {
 
 	validatorsLock       sync.RWMutex
 	validatorSyncOngoing bool
+	currentSlot          uint64
 	lastRequestedSlot    uint64
 	validatorSlotMap     map[uint64]ValidatorData
 }
@@ -161,6 +48,7 @@ func NewRemoteRelay(endpoint string, localRelay *LocalRelay) (*RemoteRelay, erro
 		client:               http.Client{Timeout: time.Second},
 		localRelay:           localRelay,
 		validatorSyncOngoing: false,
+		currentSlot:          0,
 		lastRequestedSlot:    0,
 		validatorSlotMap:     make(map[uint64]ValidatorData),
 	}
@@ -214,7 +102,11 @@ func (r *RemoteRelay) GetValidatorForSlot(nextSlot uint64) (ValidatorData, error
 	// next slot is expected to be the actual chain's next slot, not something requested by the user!
 	// if not sanitized it will force resync of validator data and possibly is a DoS vector
 
-	if r.lastRequestedSlot == 0 || nextSlot > 12+r.lastRequestedSlot {
+	r.validatorsLock.RLock()
+	defer r.validatorsLock.RUnlock()
+
+	r.currentSlot = nextSlot
+	if r.lastRequestedSlot == 0 || nextSlot/32 > r.lastRequestedSlot/32 {
 		// Every epoch request validators map
 		go func() {
 			err := r.updateValidatorsMap(nextSlot, 1)
@@ -232,9 +124,7 @@ func (r *RemoteRelay) GetValidatorForSlot(nextSlot uint64) (ValidatorData, error
 		}
 	}
 
-	r.validatorsLock.RLock()
 	vd, found := r.validatorSlotMap[nextSlot]
-	r.validatorsLock.RUnlock()
 	if found {
 		return vd, nil
 	}
@@ -242,36 +132,47 @@ func (r *RemoteRelay) GetValidatorForSlot(nextSlot uint64) (ValidatorData, error
 	return ValidatorData{}, errors.New("validator not found")
 }
 
+type BuilderSubmitBlockRequestMessage struct {
+	Slot                 uint64               `json:"slot,string"`
+	ParentHash           boostTypes.Hash      `json:"parent_hash" ssz-size:"32"`
+	BlockHash            boostTypes.Hash      `json:"block_hash" ssz-size:"32"`
+	BuilderPubkey        boostTypes.PublicKey `json:"builder_pubkey" ssz-size:"48"`
+	ProposerPubkey       boostTypes.PublicKey `json:"proposer_pubkey" ssz-size:"48"`
+	ProposerFeeRecipient boostTypes.Address   `json:"proposer_fee_recipient" ssz-size:"32"`
+	Value                boostTypes.U256Str   `json:"value" ssz-size:"32"`
+}
+
+type BuilderSubmitBlockRequest struct {
+	Signature        boostTypes.Signature             `json:"signature"`
+	Message          BuilderSubmitBlockRequestMessage `json:"message"`
+	ExecutionPayload boostTypes.ExecutionPayload      `json:"execution_payload"`
+}
+
+func (r *RemoteRelay) SubmitBlock(msg *BuilderSubmitBlockRequest) error {
+	code, err := server.SendHTTPRequest(context.TODO(), *http.DefaultClient, http.MethodPost, r.endpoint+"/relay/v1/builder/blocks", msg, nil)
+	if err != nil {
+		return err
+	}
+	if code > 299 {
+		return fmt.Errorf("non-ok response code %d from relay ", code)
+	}
+
+	if r.localRelay != nil {
+		r.localRelay.SubmitBlock(msg)
+	}
+
+	return nil
+}
+
 func (r *RemoteRelay) getSlotValidatorMapFromRelay() (map[uint64]ValidatorData, error) {
-	req, err := http.NewRequest("GET", r.endpoint+"/relay/v1/builder/validators", nil)
-	if err != nil {
-		return nil, err
-	}
-
-	req.Header.Set("accept", "application/json")
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		log.Error("client refused", "url", r.endpoint, "err", err)
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	bodyBytes, err := io.ReadAll(resp.Body)
-	if err != nil {
-		log.Error("could not read response body", "url", r.endpoint, "err", err)
-		return nil, err
-	}
-
-	if resp.StatusCode >= 300 {
-		return nil, errors.New(string(bodyBytes))
-	}
-
 	var dst GetValidatorRelayResponse
-	err = json.Unmarshal(bodyBytes, &dst)
+	code, err := server.SendHTTPRequest(context.TODO(), *http.DefaultClient, http.MethodGet, r.endpoint+"/relay/v1/builder/validators", nil, &dst)
 	if err != nil {
-		log.Error("could not unmarshal response", "url", r.endpoint, "resp", string(bodyBytes), "dst", dst, "err", err)
 		return nil, err
+	}
+
+	if code > 299 {
+		return nil, fmt.Errorf("non-ok response code %d from relay", code)
 	}
 
 	res := make(map[uint64]ValidatorData)
@@ -310,25 +211,4 @@ func (r *RemoteRelay) getSlotValidatorMapFromRelay() (map[uint64]ValidatorData, 
 	}
 
 	return res, nil
-}
-
-func (r *RemoteRelay) GetValidatorsStats() string {
-	if r.localRelay != nil {
-		return r.localRelay.GetValidatorsStats()
-	}
-
-	r.validatorsLock.RLock()
-	nValidators := len(r.validatorSlotMap)
-	r.validatorsLock.RUnlock()
-
-	return fmt.Sprint(nValidators) + " registered for current and next epochs"
-}
-
-func (r *RemoteRelay) handleRegisterValidator(w http.ResponseWriter, req *http.Request) {
-	if r.localRelay != nil {
-		r.localRelay.handleRegisterValidator(w, req)
-		return
-	}
-
-	http.Error(w, "invalid request", 400)
 }
