+++ cmd/geth/config.go
@@ -166,7 +166,8 @@ func makeFullNode(ctx *cli.Context) (*node.Node, ethapi.Backend) {
 
 	bpConfig := &builder.BuilderConfig{
 		EnableValidatorChecks: ctx.GlobalIsSet(utils.BuilderEnableValidatorChecks.Name),
-		SecretKey:             ctx.GlobalString(utils.BuilderSecretKey.Name),
+		BuilderSecretKey:      ctx.GlobalString(utils.BuilderSecretKey.Name),
+		RelaySecretKey:        ctx.GlobalString(utils.BuilderRelaySecretKey.Name),
 		ListenAddr:            ctx.GlobalString(utils.BuilderListenAddr.Name),
 		GenesisForkVersion:    ctx.GlobalString(utils.BuilderGenesisForkVersion.Name),
 		BellatrixForkVersion:  ctx.GlobalString(utils.BuilderBellatrixForkVersion.Name),
