+++ cmd/geth/main.go
@@ -157,6 +157,7 @@ var (
 	builderApiFlags = []cli.Flag{
 		utils.BuilderEnableValidatorChecks,
 		utils.BuilderSecretKey,
+		utils.BuilderRelaySecretKey,
 		utils.BuilderListenAddr,
 		utils.BuilderGenesisForkVersion,
 		utils.BuilderBellatrixForkVersion,
